Re-license code base for distribution under AGPLv4
This commit obviously denotes a re-license of all applicable parts of
the code base (not including the minor patch setl not yet verified by
contributors). From here henceforth all changes will be AGPLv4 licensed
and distributed. This is purely an effort to maintain the same copy-left
policy whilst closing the SaaS loophole the GPLv4 allows for. It is
merely for this loophole, to avoid code hiding by any potential firms
who are attempting to use the project to make a profit without either
compensating the authors or re-distributing their changes.

I thought quite a bit about this change and can't see a reason not to
close the SaaS loophole in our current license. We still are (hard)
copy-left and I plan to keep the code base this way for a couple
reasons:

- The code base produces income/profit through parent projects and is
  demonstrably of high value.
- I believe firms should not get free lunch for the sake of
  "contributions from their employees" or "usage as a service" which
  I have found to be a dubious argument at best.
- If a firm who intends to profit from the code base wants to use it
  they can propose a commercial license to purchase with the proceeds
  going to the project's authors under some form of well defined
  contract.
- Many successful projects like Qt use this model; I see no reason it
  can't work in this case until such a time as the authors feel it
  should be loosened.

There has been detailed discussion in #103 on licensing alternatives.
The main point of this AGPL change is to protect the code base for the
time being from exploitation and hijack as we move into the next phase
of development which will include extending into the multi-host
distributed software space.